chore: organize modal component files
